Move handling of MapWithDefault into NonCPS context
This is to workaround
Caused: java.io.NotSerializableException: groovy.lang.MapWithDefault

Because pipeline serializes the groovy state when calling external
functions, we can't have objects in memory which can't be serialized.
By putting such objects into a function annotated with @NonCPS, we will
never checkpoint in that function, only before or after.
That way, we can use not serializable things in such a function.