Copy over pipeline generated properties to job-dsl job
The problem with combining pipeline and job-dsl is that both tries to be
authorative about how the job should be configured.

I thought that job-dsl wouldn't touch properties that it didnt
configure, but that must have bin in older job-dsl versions.

This uses jenkins-api's to read the existing job-properties and setting
those on the new job-dsl object, to keep pipeline-generated properties.